Modify getKeys to use most pyglet keynames
Make keyDownHandler and getKeys give preference to
the use of the W3C standard keyboardEvent.code for
identifying keys pressed. The deprecated .keycode
attribute is still honored to allow older browsers
that do not support the .code attribute.

the keyList argument to getKeys should consist
of pyglet key names, allowing scripts created
using the builder to work in web experiments.