add seeding capability to psychojs random loops
I've worked out a simple way to add the seeding capability for
generating repeatable sequences of trials. It uses seedrandom()
from https://github.com/davidbau/seedrandom. This would require
adding the seedrandom.min.js from that repository to the
psychojs/js/vendors/ folder in the psychojs.zip file
that is packaged with the psychopy standalone app.

I don't know how to do that

It also requires a change to the JS_htmlHeader.tmpl in
the psychopy repository. see pull request with same title.