BF: Keep filenames in conditions file as string if starting with number
Users report issues where filenames from conditions files are not found because they begin with a number. This is because importConditions parses strings as numbers if they can be converted successfully . E.g., `01.jpg` becomes `1`.  This fix tests if a value also contains a character, and if so, do not convert to a number and preserve the string.