BF: Remove <form>...</form> elements
Leaving the elements means that a user can press enter to submit the form. Since there is no action="..." attribute, will default to submitting the form to the current page, effectively causing a page reload. Because the intention is for the user to proceed by pressing the 'Ok' button on the jQuery dialog, this only causes an annoyance to the user.